requirements.txt
pycups==2.0.1
websocket-client==1.5.3
websockets==11.0.3